additional unit-tests
