+++ integration/nwo/fsc/node/mocks/sdk1/sdk.go
@@ -1,0 +1,20 @@
+/*
+Copyright IBM Corp. All Rights Reserved.
+
+SPDX-License-Identifier: Apache-2.0
+*/
+
+package sdk1
+
+import "context"
+
+type DummySDK struct {
+}
+
+func (d *DummySDK) Install() error {
+	panic("implement me")
+}
+
+func (d *DummySDK) Start(ctx context.Context) error {
+	panic("implement me")
+}
