+++ integration/nwo/fsc/node/mocks/sdk2/sdk.go
@@ -1,0 +1,20 @@
+/*
+Copyright IBM Corp. All Rights Reserved.
+
+SPDX-License-Identifier: Apache-2.0
+*/
+
+package sdk2
+
+import "context"
+
+type DummySDK struct {
+}
+
+func (d *DummySDK) Install() error {
+	panic("implement me")
+}
+
+func (d *DummySDK) Start(ctx context.Context) error {
+	panic("implement me")
+}
